+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Mediator (0.9.4)
 
 * Removing Accounts implemented with full cleanup of associated data
+  * If admin account, correctly removes from the ADMIN list
 * Database Schema version is now recorded, allows for upgrade paths when schema changes
 * Mediator Account Type added, allows for treating the Mediator DID separately
 * FIX: Trying to strip admin rights from an empty list will now correctly create a ProblemReport that explains the issue